+++ src/Yaapii.Atoms/Enumerable/Max.cs
@@ -27,7 +27,7 @@ using Yaapii.Atoms.Enumerable;
 using Yaapii.Atoms.Fail;
 using Yaapii.Atoms.Scalar;
 
-namespace Yaapii.Atoms.List
+namespace Yaapii.Atoms.Enumerable
 {
     /// <summary>
     /// The greatest item in the given <see cref="IEnumerable{T}"/>
