+++ src/Yaapii.Atoms/Enumerable/Min.cs
@@ -27,7 +27,7 @@ using Yaapii.Atoms.Enumerable;
 using Yaapii.Atoms.Fail;
 using Yaapii.Atoms.Scalar;
 
-namespace Yaapii.Atoms.List
+namespace Yaapii.Atoms.Enumerable
 {
     /// <summary>
     /// Find the smallest item in a <see cref="IEnumerable{T}"/>
